Requested changes.

+++ Push_Back/src/math/odometry/PathPose.cpp
@@ -2,7 +2,7 @@
 
 #include "PathPose.hpp"
 
-float PathPose::distance(const PathPose& other) const {
+double PathPose::distance(const PathPose& other) const {
     return std::hypot(other.x - x, other.y - y);
 }
 
@@ -14,19 +14,19 @@ PathPose PathPose::operator-(const PathPose& other) const {
     return {x - other.x, y - other.y, velocity};
 }
 
-float PathPose::operator*(const PathPose& other) const {
+double PathPose::operator*(const PathPose& other) const {
     return x * other.x + y * other.y;
 }
 
-PathPose PathPose::operator*(const float s) const {
+PathPose PathPose::operator*(const double s) const {
     return {x * s, y * s, velocity};
 }
 
-PathPose PathPose::operator/(const float s) const {
+PathPose PathPose::operator/(const double s) const {
     return {x / s, y / s, velocity};
 }
 
-PathPose PathPose::lerp(const PathPose &other, const float t) const {
+PathPose PathPose::lerp(const PathPose &other, const double t) const {
     return {x + (other.x - x) * t, y + (other.y - y) * t, velocity};
 }
 std::string PathPose::to_string() const {
