+++ Push_Back/src/math/odometry/PathPose.hpp
@@ -32,7 +32,7 @@ public:
      * @param other The pose to get the distance to.
      * @return The distance between the two poses.
      */
-    [[nodiscard]] float distance(const PathPose& other) const;
+    [[nodiscard]] double distance(const PathPose& other) const;
 
     /**
      * Adds two poses to each-other.
@@ -60,7 +60,7 @@ public:
      * @param other
      * @return The dot product of the two poses.
      */
-    float operator*(const PathPose& other) const;
+    double operator*(const PathPose& other) const;
 
     /**
      * Multiply a pose by a scalar value.
@@ -70,7 +70,7 @@ public:
      * @param s
      * @return The pose multiplied by the scalar value.
      */
-    PathPose operator*(float s) const;
+    PathPose operator*(double s) const;
 
 
     /**
@@ -81,7 +81,7 @@ public:
      * @param s The scalar value to divide the pose by.
      * @return The pose divided by the scalar value.
      */
-    PathPose operator/(float s) const;
+    PathPose operator/(double s) const;
 
     /**
      * Linearly interpolate between two poses.
@@ -90,7 +90,7 @@ public:
      * @param t A scaling factor between 0 and 1 to how far between the two poses to interpolate.
      * @return The interpolated pose at t.
      */
-    [[nodiscard]] PathPose lerp(const PathPose& other, float t) const;
+    [[nodiscard]] PathPose lerp(const PathPose& other, double t) const;
 
     /**
      * @return Returns the string representation of the pose.
